refactor: refactor code to ES6+
